Add Migration interface and refactor SystemMigrateCasesTask
