+++ src/main/java/uk/gov/hmcts/divorce/systemupdate/schedule/SystemMigrateCasesTask.java
@@ -1,35 +1,18 @@
 package uk.gov.hmcts.divorce.systemupdate.schedule;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
-import uk.gov.hmcts.divorce.divorcecase.model.RetiredFields;
 import uk.gov.hmcts.divorce.idam.IdamService;
-import uk.gov.hmcts.divorce.systemupdate.service.CcdConflictException;
-import uk.gov.hmcts.divorce.systemupdate.service.CcdManagementException;
-import uk.gov.hmcts.divorce.systemupdate.service.CcdSearchCaseException;
-import uk.gov.hmcts.divorce.systemupdate.service.CcdSearchService;
-import uk.gov.hmcts.divorce.systemupdate.service.CcdUpdateService;
+import uk.gov.hmcts.divorce.systemupdate.schedule.migration.Migration;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.idam.client.models.User;
 
-import java.util.Map;
-
-import static org.springframework.http.HttpStatus.NOT_FOUND;
-import static uk.gov.hmcts.divorce.systemupdate.event.SystemMigrateCase.SYSTEM_MIGRATE_CASE;
+import java.util.List;
 
 @Component
 @Slf4j
 public class SystemMigrateCasesTask implements Runnable {
-
-    @Autowired
-    private CcdSearchService ccdSearchService;
-
-    @Autowired
-    private CcdUpdateService ccdUpdateService;
 
     @Autowired
     private IdamService idamService;
@@ -38,7 +21,7 @@ public class SystemMigrateCasesTask implements Runnable {
     private AuthTokenGenerator authTokenGenerator;
 
     @Autowired
-    private ObjectMapper objectMapper;
+    private List<Migration> migrations;
 
     @Override
     public void run() {
@@ -47,95 +30,6 @@ public class SystemMigrateCasesTask implements Runnable {
         final User user = idamService.retrieveSystemUpdateUserDetails();
         final String serviceAuthorization = authTokenGenerator.generate();
 
-        baseMigration(user, serviceAuthorization);
-
-        migrateJointApplications(user, serviceAuthorization);
-    }
-
-    private void migrateJointApplications(User user, String serviceAuthorization) {
-        log.info("Started migrating cases joint application");
-        try {
-            ccdSearchService
-                .searchJointApplicationsWithAccessCodePostIssueApplication(user, serviceAuthorization)
-                .parallelStream()
-                .forEach(details -> removeAccessCode(details, user, serviceAuthorization));
-
-        } catch (final CcdSearchCaseException e) {
-            log.error("Case schedule task(migration joint application) stopped after search error", e);
-        }
-    }
-
-
-    private void baseMigration(User user, String serviceAuthorization) {
-        try {
-            ccdSearchService
-                .searchForCasesWithVersionLessThan(RetiredFields.getVersion(), user, serviceAuthorization)
-                .parallelStream()
-                .forEach(details -> migrateCase(details, user, serviceAuthorization));
-
-        } catch (final CcdSearchCaseException e) {
-            log.error("Case migration schedule task stopped after search error", e);
-        }
-    }
-
-    private void migrateCase(final CaseDetails caseDetails, final User user, final String serviceAuthorization) {
-        final Long caseId = caseDetails.getId();
-
-        try {
-            final var data = RetiredFields.migrate(caseDetails.getData());
-            verifyData(data, caseId);
-
-            caseDetails.setData(data);
-            ccdUpdateService.submitEvent(caseDetails, SYSTEM_MIGRATE_CASE, user, serviceAuthorization);
-            log.info("Migration complete for case id: {}", caseId);
-        } catch (final CcdConflictException e) {
-            log.error("Could not get lock for case id: {}, continuing to next case", caseId);
-        } catch (final CcdManagementException e) {
-            log.error("Submit event failed for case id: {}, continuing to next case", caseId);
-            failedMigrationSetVersionToZero(caseDetails, user, serviceAuthorization, caseId, e);
-        }
-    }
-
-    private void removeAccessCode(final CaseDetails caseDetails, final User user, final String serviceAuthorization) {
-        final Long caseId = caseDetails.getId();
-
-        try {
-            caseDetails.getData().put("accessCode", "");
-            ccdUpdateService.submitEvent(caseDetails, SYSTEM_MIGRATE_CASE, user, serviceAuthorization);
-            log.info("Removed access code successfully for case id: {}", caseId);
-        } catch (final CcdConflictException e) {
-            log.error("Could not get lock for case id: {}, continuing to next case", caseId);
-        } catch (final CcdManagementException e) {
-            log.error("Submit event(after removing access code) failed for case id: {}, continuing to next case", caseId);
-        }
-    }
-
-    private void failedMigrationSetVersionToZero(final CaseDetails caseDetails,
-                                                 final User user,
-                                                 final String serviceAuthorization,
-                                                 final Long caseId,
-                                                 final CcdManagementException ccdManagementException) {
-
-        if (ccdManagementException.getStatus() != NOT_FOUND.value()) {
-            log.info("Setting dataVersion to 0 for case id: {} after failed migration", caseDetails.getId());
-
-            caseDetails.setData(Map.of("dataVersion", 0));
-            ccdUpdateService.submitEvent(caseDetails, SYSTEM_MIGRATE_CASE, user, serviceAuthorization);
-
-            log.info("dataVersion set for case id: {}", caseDetails.getId());
-        } else {
-            log.info("Version not set to 0 case not found for case id: {}", caseId);
-        }
-    }
-
-    private void verifyData(Map<String, Object> data, Long id) {
-        try {
-            objectMapper.convertValue(data, CaseData.class);
-        } catch (final IllegalArgumentException e) {
-            log.info("Migration failed for case id {} due to deserialization error", id);
-            log.info("Deserialization error caused by {}", e.getMessage());
-
-            data.put("dataVersion", 0);
-        }
+        migrations.forEach(migration -> migration.apply(user, serviceAuthorization));
     }
 }
